requirements.txt
Flask==3.1.0
Flask-Caching==1.10.1
google-cloud-datastore==2.20.1
google-cloud-bigquery==3.27.0
pytest==8.3.4
requests==2.32.3
